Updated for FW-1.0.2
Firmware Update 1.0.2 made a few changes to some of the base module pages. To include but limited to defaulting the # of Clients and SSIDs on the Dashboard to be Black font color through the module HTML and Linking them through href so that you can more easily to the associated pages. I have updated the CSS to allow for the #'s to be seen without having to go in and edit the HTML pages for the modules.

I have also added CSS to the Login page that brings the entirety of the format together to include having the Login button be red instead of green and the outline shadows on the Username and Password maintain a red frame while selected.